Renamed src/components/burnplanning to [...]burn-planning
